Refactor CTA container class
I followed Claude's advice to use min() on the
container which is supposedly more modern and
robust. It broke the CTA container. I fixed the CTA
container by removing the width so it followed
the container width calculation. I also re-ordered
the properties to follow Ryan from CodeStitch's
style guide.