Add mobile styles to hero container
It bothers me that the homepage isn't mobile
responsive. I put Claude's suggestions for css
rules using media-queries. It looks OK for the
smaller screens, i.e., it's mobile responsive.
But it broke the desktop site. The large
desktop version of the site has too many
elements that are huge and stack.